Update dependencies and enhance grid information handling
- Bump c-two dependency version to 0.1.17 in pyproject.toml and uv.lock.
- Refactor activate_grid_info endpoint to return binary data for active grid information.
- Simplify get_active_grid_infos function in grid_comp.py.
- Introduce MultiGridInfo class in grid.py for structured grid data handling.
- Add error handling in MultiGridInfo fetching methods.
- Remove unused code and improve overall code organization.
