+++ pyproject.toml
@@ -5,7 +5,7 @@ description = "Add your description here"
 readme = "README.md"
 requires-python = ">=3.12.10"
 dependencies = [
-    "c-two>=0.1.16",
+    "c-two>=0.1.17",
     "fastapi>=0.115.12",
     "gdal==3.10.2",
     "jinja2>=3.1.6",
